pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-xkit"
version = "0.2.0"
description = "FastAPI production toolkit: Redis rate limiting, JWT auth, RBAC, audit logs to DB, OpenTelemetry middleware."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name="FastAPI-XKit Contributors"}]
keywords = ["fastapi", "middleware", "rate-limit", "redis", "jwt", "rbac", "audit", "opentelemetry"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Framework :: FastAPI",
  "Intended Audience :: Developers",
  "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
  "fastapi>=0.100.0",
  "starlette>=0.27.0",
  "pydantic>=2.0.0",
  "uvicorn>=0.23.0",
  "PyJWT>=2.8.0",
  "redis>=5.0.0",
  "sqlalchemy>=2.0.0",
  "opentelemetry-api>=1.22.0",
  "opentelemetry-sdk>=1.22.0",
  "opentelemetry-instrumentation-asgi>=0.43b0",
  "opentelemetry-exporter-otlp>=1.22.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "httpx>=0.24.0",
  "build>=1.0.0",
  "twine>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/yourname/fastapi-xkit"
Repository = "https://github.com/yourname/fastapi-xkit"

[tool.setuptools.packages.find]
where = ["."]
include = ["fastapixkit*"]
